UI class: update methods
- Overload current methods for printing expenditure actions to allow Entry object to be passed directly
- Document current methods
- Add printException to allow formatted output of exception messages
- Refactor helper methods which do not directly interact with the UI to UIUtil.java
